Quote ticker render function
With this change we're adding a render() function to the QuoteTicker
view. Every view has a function called render() which is responsible for
modifying the page DOM to have the most up-to-date content for its view.

Within the render() function, `this.$el` refers to the DOM element that
this particular view instance is "attached" to. It is best practice to
not modify anything about the page other than that element and its
children.

With the render() function for QuoteTicker we are setting the contents
of the attached element to be whatever is rendered by the previously
established template, when we pass in a JavaScript object with the
necessary variables.

By convention, the render() function always ends with `return this;`.