Display quote ticker within application view
In this change we are rewriting the Application view's render() function
to use the new QuoteTicker view that we just wrote.

The render() function creates a new instance of QuoteTicker, and passes
into the constructor function a JavaScript object with customization or
other instance-specific data.

In our case this is the `el` property, which tells the QuoteTicker view
what DOM element to "attach" to. For now, this is just the element that
the Application view itself is attached to.

Next we tell the QuoteTicker instance to render itself, calling the code
we wrote in the previous commit.

Finally, we `return this;` as we should always do in a render()
function.