Refactor QuoteTicker to not use hardcoded data
In our original version of the QuoteTicker's render() function we would
always display the same stock symbol and current price for every
instance. This doesn't make a ton of sense, as we want to have different
tickers for different stocks!

In order to make that feasible, in this change we are refactoring our
code a bit to use an instance-specific property passed into the
QuoteTicker constructor function.

By convention in Backbone.js the JavaScript object holding the variables
used when constructing a view is called the model. There is a strong
correlation between the concept of a model in both Backbone.js and
Rails.

We will see later on that Backbone.js has its own way of creating model
"classes", which have additional features we can take advantage of.
However for now, we will be using a Plain Old JavaScript Object (POJO)
as our model.

In order to make that work, we have our Application view pass in a new
property when it creates the QuoteTicker instance: `model`. This
property has a value which is exactly the same as the one we used in
QuoteTicker's render() function previously.

Now instead of using a literal object in QuoteTicker's render()
function we can reference the model property, which is available on
`this`.