Application view support for multiple tickers
In this change we are updating the Application view's render() function
to have better support for displaying multiple QuoteTicker instances.

Right now our Application's render() creates a single QuoteTicker
instance and attaches it directly to the same element the Application
itself is attached to. This can't work for multiple QuoteTickers as they
would be "fighting" for whose ticker content would get displayed in that
space.

To prevent any tickers from overwriting each other, we instead create a
new <div> for the ticker and append it into the element the Application
is attached to.

The single QuoteTicker that we currently have is then created with this
new <div> as its attachment point. When we add more QuoteTickers they
can each have their own <div> and will play nicely with each other.