Refactor Application ticker rendering
In this change we're doing a small refactor to better support multiple
tickers when we add them in the future.

The refactor consists of moving the majority of the contents of
Application's render() function into a new function named renderTicker.

This new function will be responsible for rendering a single ticker into
the Application's attachment point when given the appropriate data for
that particular ticker.

The only change between the contents of renderTicker() and the previous
render() function are that the POJO we are passing to QuoteTicker as the
model property is turned into a variable within renderTicker().

The new render() function calls renderTicker() and passes to it the
original contents of the POJO for QuoteTicker's model.