Price change event on quote ticker
This change adds an event to the QuoteTicker view and immediately
triggers it.

Backbone.js events are similar to jQuery events, except they're
generally more flexible. Backbone.js events do not need to relate to
any particular user action (such as clicking).

Instead, we define our own events by setting up an event handler and
then have other parts of our code initiate the event by "triggering" or
"firing" the event (along with any relevant parameters).

In this case we're establishing a "price-change" event for QuoteTicker
with the call to `this.on()`. The use of `this` is because the object
that will receive the event notification is the QuoteTicker itself.

When we setup the event with on() we provide a function which will act
as the event handler. In this case we've written a short handler called
onPriceChange() which takes a priceChange parameter and updates the model
by adding it to the current price.

The onPriceChange() handler also calls the render() function so that our
view is updated properly. If we did not call render() the price in the
model would change, but we would see no difference on the page --
because we did not run any jQuery code to change the DOM.

Finally, in the initialize() method after setting up the "price-change"
event we immediately trigger it using `this.trigger()` and giving a
change value of +100 dollars. This is called "money printing".