Stock quotes move randomly
In this change we show a more realistic use of the "price-change" event
by triggering it once per second.

Frankly, the previous commit contained some highly unrealistic market
behavior! All stock prices went up $100 from their starting price which
I'm sure was met with immediate elation by the investing public,
followed shortly thereafter by a sudden realization that this degree of
inflation has likely just made everyone poorer. Bummer.

Instead, we must provide our ever-gullible investors with something the
Wall Street types refer to as "price action". In layman's terms the
price of our stocks must frequently move up and down.

This provides fertile ground for the investors' minds to run wild and
apply all manner of elaborate logic in an effort to establish meaning
for what is, ultimately, a completely random[1] process.

It also gives the smartest[2] investors an opportunity to "earn" a
profit through the judicious application of buy and sell orders (once
those are implemented). Exciting!

The code for how we calculate a price movement is not particularly
important here. The main thing to pay attention to is the use of
setInterval() to run the given anonymous function once every 1,000
milliseconds (1 Hz). The anonymous function then triggers the
"price-change" event, as we did in the previous commit.

[1] - pseudorandom
[2] - luckiest