Buy and Sell buttons on Application
These buttons are static page elements so we're adding them directly
into the index.html file.

We're placing them inside of the <div> where the Application view is
attached, so that they are more easily accessible to it for the event
logic which is coming next.

Placing them inside of the attachment point for Application also keeps
to the rule that your view code should never interact directly with any
page elements that are not inside of the element it's attached to.