Keep track of all quote tickers on Application
Previously we were creating QuoteTicker view instances when rendering
the Application, but we were not holding on to those instances.

While they remained around so that our event logic could continue
updating the prices, after rendering they were no longer accessible to
the Application view.

We now push them into a `tickers` array on the Application view for us
to later use.