Buy button logic for Application view
This change establishes an event handler on the Application view for
dealing with the click event on our buy button.

Previously we have seen how to setup an event handler where the "target"
of the event was a Backbone object (such as the QuoteTicker view). In
this case we're handling an event that comes from user interaction with
a page element -- clicking on a button -- so the code to setup the event
handler is different.

For a view that has interactable page elements, we can setup an `events`
property in the view class definition. This property is a plain
JavaScript object with keys that define the event and targets, and
values that identify a handler function.

The keys use the format "<event type> <target select>". For example,
"click .btn-buy" establishes an event handler for the "click" event on
any element (within the view's attachment point) that has the class
"btn-buy".

By setting the associated value for the event key to "onClickBuy",
Backbone.js knows to run the function named onClickBuy() defined on the
current view. This function works in the same way as the event handlers
we used in jQuery (in fact this system works by using jQuery), so the
event handler function is given an `event` object with all of the
details of what event occurred.

Our event handler named onClickBuy() loops through every QuoteTicker in
the tickers array and triggers the "price-change" event on it. The
handler passes in a price increase of $1 which represents the impact of
our buying every stock. This is known as "diversification".