Sell button logic for Application view
This change adds the event handler for the sell button that is part of
the Application view. Its behavior is similar to the buy button, but
with an opposite price movement.

Be careful with this sell button! If you sell too often the market will
tank and the exchanges will have to step in and halt trading through a
mechanism known as a "circuit breaker". This is bad and makes the Wall
Street types very sad.

This is what happens when your market truly breaks free and runs like
the wind (straight into the ground).