Refactor Application view to use Application model
In this change we begin the transition to using models for all of our
data, rather than plain JavaScript objects that live inside of our
views.

The reason for using models to store our data in Backbone.js is much the
same was it was in Rails. When we separate our concerns the code for
each separate piece becomes easier to understand, easier to test, and
more flexible in how it can be combined with other code.

In this change we're using the existing Application model that came with
the project baseline, but until now has been empty.

For this refactor we're moving the quote information out of the
Application view and onto the model as a property named `quotes`. This
property is an array because we have two pieces of quote information.

In order to access the information that is now stored in the Application
model, we must use `this.model` within our Application view. The reason
that `this.model` works is because the `model` property on the view was
provided during its creation, in the `src/app.js` file.

Since the `quotes` property on our Application model is an array, we now
must loop through it with `forEach` and call the same renderTicker()
function with each piece of quote data.