Quote model
In this change we define a new model named Quote by extending
`Backbone.Model` and providing some default attribute values.

The syntax for defining a new model is very similar to that for defining
a new Backbone view, the main difference being that we use
`Backbone.Model` instead of `Backbone.View`.

While our model and view classes can have similar functionality in some
cases, such as the initialize() function, they are mostly different. For
example, views always have a render() function and `el` property, and
usually have a `model` property.

By contrast, our model classes have things like the `defaults` property
which defines a set of default attributes (and their values) that each
instance of the model starts off with.

Model classes will also frequently have functions on them that are
specific to the class, like a buy() function for a stock quote.

Finally, Backbone.js models include code to support using them with a
RESTful API to allow for easy sending & retreiving from a back-end
server.

A Note About Properties & Attributes:
It may seem like I have been using these terms interchangeably, however
they have very distinct meanings in Backbone.js.

Properties are instance-level variables that can be defined directly on
a Backbone class when it is defined (with `Backbone.Whatever.extend()`).
In the case of views, additional properties can be set on a single
instance by passing them in when the instance is created (with
`new MyView()`).

In both cases properties are supposed to be static variables that we set
once and don't update later. The contents of the model that a view uses
may change, but the `model` property will always "point" to the same
model instance. All of the functions that we've defined, such as
render() and onClickBuy(), are properties.

Attributes are specific to model classes, and they represent the
instance-level variables that are the non-static data for that instance.
They are most similar to the attributes on Rails models that are created
with database migrations, and that are stored in database tables.

When you create a new instance of a model you can pass in a JavaScript
object that contains attributes (and their values) for that particular
instance, which will override any default set with the same name.

The attributes of a model cannot be accessed in the same way as the
properties (with `someModel.property`), but instead should be used
through the built-in get() and set() functions.

Calling `someModel.set("property", "value")` will update the attribute
and possibly trigger events that indicate to the rest of our code that
this model has changed.