Use Quote model (part 1)
In this change the Application model is updated to use its static quote
data property to create an attribute that is a list of Quote instances.

We've renamed the `quotes` property to `quotesData` as this is more
accurate to its new purpose. In the new initialize() method, we add a
new attribute named `quotes`, which we set() to be an array of Quote
instances by using Underscore's `_.map()` function.