Use Quote model (part 2)
In this change we update the QuoteTicker view to work with a Quote
instance as its `model` property.

First, when running the compiled Underscore template function we cannot
pass the Quote instance directly. Instead we pass in the `attributes`
property of the instance, which is a JavaScript object that includes all
of the attributes from the model (`symbol` and `price` in this case).

Next, we have the onPriceChange() event handler defer to the Quote
instance to update its own price. This is the first step in separating
out the "business" logic from the view logic and placing it in the model
instead.

The priceChange() function on the Quote model is equivalent to the
original code from onPriceChange(), however because the Quote model has
its price as an attribute we need to use the `get()` and `set()` methods
to implement the accumulation.