Move price action logic into Quote model (part 1)
In this change we move the logic for randomly updating a quote's price
into the Quote model, and have the QuoteTicker view instead listen to
the Quote for any changes.

This setup is more or less the recommended approach for Backbone's MVC
style. The model code has all of the logic necesary to change the data,
and the view code is setup to "listen" for events triggered by the
model.

When a model event is triggered that the view is listening for, it runs
an event handler function which, more often than not, renders the view
again with the updated data from the model.

That's exactly what we're doing in this case, with the QuoteTicker view
listening to the Quote model for any "change" events. When a "change"
even comes in from the model, we run an event handler function which
does nothing but re-render the view.

In every Backbone model class, when an attribute is changed by calling
the set() function, the "change" event is triggered. This means that the
newly-moved random price action logic which calls priceChange()
ultimately results in a "change" event, which then results in the
QuoteTicker view being re-rendered.