Move price action logic into Quote model (part 2)
In this change we fix up the Application view's Buy and Sell buttons to
work with the Quote instances that are on the Application model.

Since we took out the event handler for the "price-change" event on the
QuoteTicker view... the Application view's event handlers cannot trigger
them anymore.

Instead, this view should talk directly to the Quote instances that it
can get() from it associated Application model. When we get the Quote
instances we loop through them with forEach() as before, and call a new
function buy() or sell() on each one, depending on which button was
clicked.

The buy() and sell() functions are written directly on the Quote model,
and represent additional actions that we can take for a given Quote.

They both have equivalent results as their original implementation with
the "price-change" event trigger, but now the logic for how the price
changes when we buy or sell a particular stock is isolated entirely
within the Quote model.