Buy and sell individual stocks
In this change we add Buy and Sell buttons to each QuoteTicker which
allow the user to buy or sell an individual stock.

Because of the work we've done previously with refactoring and moving
our logic into the models, the code to implement this functionality is
very short.

We set up an `events` property on the QuoteTicker, similar to the one
that we set up on the Application view previously. The event handlers
for clicking on our two buttons only need to call buy() or sell() on the
Quote model instance associated with this view in order to complete
their work.

Now our investors can implement their most advanced trading techniques
yet! Buy one stock and sell another. Buy two stocks then sell two
stocks. Sell their vast borrowings of a single stock and watch as the
market panics and races to follow your "momentum", then buy it all back
again at rock-bottom prices from all the suckers looking to hold onto
the last miniscule scraps of their retirement.

This is known as a "bear raid" because when you do it the SEC sends
bears to raid your home. Just kidding, the SEC doesn't prosecute anyone.