Trade model
In this change we add the basic structure for a Trade model which will
be kept in a TradeLog collection.

The Trade model represents the historical record of a trade that took
place in the past. Instances of the Trade model have attributes for the
symbol and quantity that was traded, the price it was traded at, the
date of the trade, and whether the trade resulted from a buy or a sell
order.