TradeLog collection
In this change we add the core structure for a TradeLog collection
class, that represents a sequence of Trade model instances.

Every Backbone collection class is associated with a particular Backbone
model class, defined through the `model` property. Collection classes
come built-in with all of the Underscore functionality for arrays, and
support for working with RESTful APIs. They also trigger events whenever
an element is added or removed from them.